feat(cloudflare): support cross-binding to Workflow
