+++ alchemy-web/docs/guides/cloudflare-vitejs.md
@@ -1,5 +1,5 @@
 ---
-order: 1
+order: 1.5
 title: Vite.js
 description: Step-by-step guide to deploying Vite.js React applications with API endpoints to Cloudflare Workers using Alchemy's Infrastructure-as-Code approach.
 ---
