+++ alchemy/test/cloudflare/d1-database.test.ts
@@ -2,8 +2,10 @@ import { describe, expect } from "bun:test";
 import { alchemy } from "../../src/alchemy.js";
 import { createCloudflareApi } from "../../src/cloudflare/api.js";
 import { D1Database, listDatabases } from "../../src/cloudflare/d1-database.js";
+import { Worker } from "../../src/cloudflare/worker.js";
 import { BRANCH_PREFIX } from "../util.js";
 
+import { destroy } from "../../src/destroy.js";
 import "../../src/test/bun.js";
 
 const test = alchemy.test(import.meta, {
@@ -357,6 +359,119 @@ describe("D1 Database Resource", async () => {
       await alchemy.destroy(scope);
     }
   });
+
+  test("create and test worker with D1 database binding", async (scope) => {
+    const workerName = `${BRANCH_PREFIX}-test-worker-d1`;
+
+    let worker: Worker<{ DATABASE: D1Database }> | undefined;
+    let db: D1Database | undefined;
+
+    try {
+      // Create a D1 database
+      db = await D1Database(`${BRANCH_PREFIX}-test-db`, {
+        name: `${BRANCH_PREFIX}-test-db`,
+        primaryLocationHint: "wnam", // West North America
+      });
+
+      expect(db.id).toBeTruthy();
+      expect(db.name).toEqual(`${BRANCH_PREFIX}-test-db`);
+
+      // Create a worker with the D1 database binding
+      worker = await Worker(workerName, {
+        name: workerName,
+        script: `
+          export default {
+            async fetch(request, env, ctx) {
+              const url = new URL(request.url);
+
+              // Initialize the database with a table and data
+              if (url.pathname === '/init-db') {
+                try {
+                  const db = env.DATABASE;
+
+                  // Create a test table
+                  await db.exec("CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY, name TEXT, email TEXT)");
+
+                  // Insert some test data
+                  await db.exec("INSERT INTO users (name, email) VALUES ('Test User', 'test@example.com')");
+
+                  return new Response('Database initialized successfully!', {
+                    status: 200,
+                    headers: { 'Content-Type': 'text/plain' }
+                  });
+                } catch (error) {
+                  return new Response('Error initializing database: ' + error.message, {
+                    status: 500,
+                    headers: { 'Content-Type': 'text/plain' }
+                  });
+                }
+              }
+
+              // Query data from the database
+              if (url.pathname === '/query-db') {
+                try {
+                  const db = env.DATABASE;
+
+                  // Query the database
+                  const { results } = await db.prepare("SELECT * FROM users").all();
+
+                  return new Response(JSON.stringify({ success: true, data: results }), {
+                    status: 200,
+                    headers: { 'Content-Type': 'application/json' }
+                  });
+                } catch (error) {
+                  return new Response(JSON.stringify({
+                    success: false,
+                    error: error.message
+                  }), {
+                    status: 500,
+                    headers: { 'Content-Type': 'application/json' }
+                  });
+                }
+              }
+
+              return new Response('D1 Database Worker is running!', {
+                status: 200,
+                headers: { 'Content-Type': 'text/plain' }
+              });
+            }
+          };
+        `,
+        format: "esm",
+        url: true, // Enable workers.dev URL to test the worker
+        bindings: {
+          DATABASE: db,
+        },
+      });
+
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+
+      expect(worker.id).toBeTruthy();
+      expect(worker.name).toEqual(workerName);
+      expect(worker.bindings).toBeDefined();
+      expect(worker.bindings!.DATABASE).toBeDefined();
+      expect(worker.bindings!.DATABASE.id).toEqual(db.id);
+      expect(worker.url).toBeTruthy();
+
+      // Initialize the database with a table and data
+      const initResponse = await fetch(`${worker.url}/init-db`);
+      expect(initResponse.status).toEqual(200);
+      const initText = await initResponse.text();
+      expect(initText).toEqual("Database initialized successfully!");
+
+      // Query data from the database
+      const queryResponse = await fetch(`${worker.url}/query-db`);
+      expect(queryResponse.status).toEqual(200);
+      const queryData: any = await queryResponse.json();
+      expect(queryData.success).toEqual(true);
+      expect(queryData.data).toBeArray();
+      expect(queryData.data.length).toBeGreaterThan(0);
+      expect(queryData.data[0].name).toEqual("Test User");
+      expect(queryData.data[0].email).toEqual("test@example.com");
+    } finally {
+      await destroy(scope);
+    }
+  }, 120000); // Increased timeout for D1 database operations
 });
 
 async function assertDatabaseDeleted(database: D1Database) {
