+++ alchemy/test/cloudflare/queue.test.ts
@@ -2,6 +2,8 @@ import { describe, expect } from "bun:test";
 import { alchemy } from "../../src/alchemy.js";
 import { createCloudflareApi } from "../../src/cloudflare/api.js";
 import { Queue, listQueues } from "../../src/cloudflare/queue.js";
+import { Worker } from "../../src/cloudflare/worker.js";
+import { destroy } from "../../src/destroy.js";
 import { BRANCH_PREFIX } from "../util.js";
 
 import "../../src/test/bun.js";
@@ -173,6 +175,110 @@ describe("Cloudflare Queue Resource", async () => {
       await alchemy.destroy(scope);
     }
   }, 120000);
+
+  test("create and test worker with Queue binding", async (scope) => {
+    // Sample ESM worker script with Queue functionality
+
+    const workerName = `${BRANCH_PREFIX}-test-worker-queue`;
+    const queueName = `${BRANCH_PREFIX}-test-queue`;
+
+    let worker: Worker<{ MESSAGE_QUEUE: Queue }> | undefined;
+    let queue: Queue | undefined;
+
+    try {
+      // Create a Queue
+      queue = await Queue(queueName, {
+        name: queueName,
+        settings: {
+          deliveryDelay: 0, // No delay for testing
+          deliveryPaused: false,
+        },
+      });
+
+      expect(queue.id).toBeTruthy();
+      expect(queue.name).toEqual(queueName);
+      expect(queue.type).toEqual("queue");
+
+      // Create a worker with the Queue binding
+      worker = await Worker(workerName, {
+        name: workerName,
+        script: `
+          export default {
+            async fetch(request, env, ctx) {
+              const url = new URL(request.url);
+
+              // Send a message to the queue
+              if (url.pathname === '/send-message') {
+                try {
+                  const body = await request.json();
+                  const messageId = await env.MESSAGE_QUEUE.send(body);
+
+                  return new Response(JSON.stringify({
+                    success: true,
+                    messageId,
+                    message: 'Message sent successfully'
+                  }), {
+                    status: 200,
+                    headers: { 'Content-Type': 'application/json' }
+                  });
+                } catch (error) {
+                  return new Response(JSON.stringify({
+                    success: false,
+                    error: error.message
+                  }), {
+                    status: 500,
+                    headers: { 'Content-Type': 'application/json' }
+                  });
+                }
+              }
+
+              return new Response('Queue Worker is running!', {
+                status: 200,
+                headers: { 'Content-Type': 'text/plain' }
+              });
+            }
+          };
+        `,
+        format: "esm",
+        url: true, // Enable workers.dev URL to test the worker
+        bindings: {
+          MESSAGE_QUEUE: queue,
+        },
+      });
+
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+
+      expect(worker.id).toBeTruthy();
+      expect(worker.name).toEqual(workerName);
+      expect(worker.bindings).toBeDefined();
+      expect(worker.bindings!.MESSAGE_QUEUE).toBeDefined();
+      expect(worker.url).toBeTruthy();
+
+      if (worker.url) {
+        // Send a message to the queue
+        const testMessage = {
+          id: "msg-123",
+          content: "Test message content",
+          timestamp: Date.now(),
+        };
+
+        const sendResponse = await fetch(`${worker.url}/send-message`, {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify(testMessage),
+        });
+
+        expect(sendResponse.status).toEqual(200);
+        const responseData: any = await sendResponse.json();
+        expect(responseData.success).toEqual(true);
+        expect(responseData.message).toEqual("Message sent successfully");
+      }
+    } finally {
+      await destroy(scope);
+    }
+  }, 120000); // Increased timeout for Queue operations
 });
 
 async function assertQueueDeleted(queue: Queue) {
