+++ alchemy/test/cloudflare/worker.test.ts
@@ -6,13 +6,9 @@ import { AnalyticsEngineDataset } from "../../src/cloudflare/analytics-engine.js
 import { createCloudflareApi } from "../../src/cloudflare/api.js";
 import { Assets } from "../../src/cloudflare/assets.js";
 import { Self } from "../../src/cloudflare/bindings.js";
-import { R2Bucket } from "../../src/cloudflare/bucket.js";
-import { D1Database } from "../../src/cloudflare/d1-database.js";
 import { DurableObjectNamespace } from "../../src/cloudflare/durable-object-namespace.js";
 import { KVNamespace } from "../../src/cloudflare/kv-namespace.js";
-import { Queue } from "../../src/cloudflare/queue.js";
 import { Worker } from "../../src/cloudflare/worker.js";
-import { Workflow } from "../../src/cloudflare/workflow.js";
 import { destroy } from "../../src/destroy.js";
 import { BRANCH_PREFIX } from "../util.js";
 
@@ -39,229 +35,6 @@ async function assertWorkerDoesNotExist(workerName: string) {
 }
 
 describe("Worker Resource", () => {
-  // Sample worker script (CJS style)
-  const workerScript = `
-    addEventListener('fetch', event => {
-      event.respondWith(new Response('Hello world!', { status: 200 }));
-    });
-  `;
-
-  // Sample ESM worker script
-  const esmWorkerScript = `
-    export default {
-      async fetch(request, env, ctx) {
-        return new Response('Hello ESM world!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample ESM worker script with a Durable Object
-  const durableObjectWorkerScript = `
-    export class Counter {
-      constructor(state, env) {
-        this.state = state;
-        this.env = env;
-        this.counter = 0;
-      }
-
-      async fetch(request) {
-        this.counter++;
-        return new Response('Counter: ' + this.counter, { status: 200 });
-      }
-    }
-
-    export default {
-      async fetch(request, env, ctx) {
-        // Use the DO binding if needed
-        if (request.url.includes('/counter')) {
-          const id = env.COUNTER.idFromName('default');
-          const stub = env.COUNTER.get(id);
-          return stub.fetch(request);
-        }
-
-        return new Response('Hello with Durable Object!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample ESM worker script with KV Namespace
-  const kvWorkerScript = `
-    export default {
-      async fetch(request, env, ctx) {
-        // Use the KV binding
-        if (request.url.includes('/kv')) {
-          const value = await env.TEST_KV.get('testKey');
-          return new Response('KV Value: ' + (value || 'not found'), { status: 200 });
-        }
-
-        return new Response('Hello with KV Namespace!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample ESM worker script with R2 bucket
-  const r2WorkerScript = `
-    export default {
-      async fetch(request, env, ctx) {
-        // Use the R2 binding
-        if (request.url.includes('/r2-info')) {
-          // Just confirm we have access to the binding
-          return new Response(JSON.stringify({
-            hasR2: !!env.STORAGE,
-            bucketName: env.STORAGE.name || 'unknown'
-          }), {
-            status: 200,
-            headers: { 'Content-Type': 'application/json' }
-          });
-        }
-
-        return new Response('Hello with R2 Bucket!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample ESM worker script with multiple bindings
-  const multiBindingsWorkerScript = `
-    export class Counter {
-      constructor(state, env) {
-        this.state = state;
-        this.env = env;
-        this.counter = 0;
-      }
-
-      async fetch(request) {
-        this.counter++;
-        return new Response('Counter: ' + this.counter, { status: 200 });
-      }
-    }
-
-    export default {
-      async fetch(request, env, ctx) {
-        // Path-based routing to demonstrate different bindings
-        const url = new URL(request.url);
-
-        if (url.pathname.includes('/counter')) {
-          const id = env.COUNTER.idFromName('default');
-          const stub = env.COUNTER.get(id);
-          return stub.fetch(request);
-        }
-
-        if (url.pathname.includes('/kv')) {
-          const value = await env.TEST_KV.get('testKey');
-          return new Response('KV Value: ' + (value || 'not found'), { status: 200 });
-        }
-
-        if (url.pathname.includes('/secret')) {
-          return new Response('Secret: ' + env.API_KEY, { status: 200 });
-        }
-
-        return new Response('Hello worker with multiple bindings!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample ESM worker script with environment variables
-  const envVarsWorkerScript = `
-    export default {
-      async fetch(request, env, ctx) {
-        const url = new URL(request.url);
-
-        // Return the value of the requested environment variable
-        if (url.pathname.startsWith('/env/')) {
-          const varName = url.pathname.split('/env/')[1];
-          const value = env[varName];
-          return new Response(value || 'undefined', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-
-        // Return all environment variables
-        if (url.pathname === '/env') {
-          const envVars = Object.entries(env)
-            .filter(([key]) => key !== 'COUNTER' && !key.includes('Durable')) // Filter out bindings
-            .map(([key, value]) => \`\${key}: \${value}\`)
-            .join('\\n');
-
-          return new Response(envVars, {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-
-        return new Response('Hello with environment variables!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample ESM worker script with original Counter class
-  const doMigrationWorkerScriptV1 = `
-    export class Counter {
-      constructor(state, env) {
-        this.state = state;
-        this.env = env;
-        this.counter = 0;
-      }
-
-      async fetch(request) {
-        this.counter++;
-        return new Response('Counter V1: ' + this.counter, { status: 200 });
-      }
-    }
-
-    export default {
-      async fetch(request, env, ctx) {
-        if (request.url.includes('/counter')) {
-          const id = env.COUNTER.idFromName('default');
-          const stub = env.COUNTER.get(id);
-          return stub.fetch(request);
-        }
-        return new Response('Hello with Counter V1!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample ESM worker script with renamed CounterV2 class
-  const doMigrationWorkerScriptV2 = `
-    export class CounterV2 {
-      constructor(state, env) {
-        this.state = state;
-        this.env = env;
-        this.counter = 0;
-      }
-
-      async fetch(request) {
-        this.counter++;
-        return new Response('Counter V2: ' + this.counter, { status: 200 });
-      }
-    }
-
-    export default {
-      async fetch(request, env, ctx) {
-        if (request.url.includes('/counter')) {
-          const id = env.COUNTER.idFromName('default');
-          const stub = env.COUNTER.get(id);
-          return stub.fetch(request);
-        }
-        return new Response('Hello with Counter V2!', { status: 200 });
-      }
-    };
-  `;
-
-  // Sample worker script with a scheduled handler
-  const cronWorkerScript = `
-    export default {
-      async fetch(request, env, ctx) {
-        return new Response('Worker with cron is running!', { status: 200 });
-      },
-      async scheduled(event, env, ctx) {
-        // Log the scheduled event details
-        console.log('Scheduled event received:', event.scheduledTime, event.cron);
-        // In a real worker, you would perform tasks here
-      },
-    };
-  `;
-
   test("create, update, and delete worker (CJS format)", async (scope) => {
     const workerName = `${BRANCH_PREFIX}-test-worker-cjs-1`;
 
@@ -270,7 +43,11 @@ describe("Worker Resource", () => {
       // Create a worker with an explicit name
       worker = await Worker(workerName, {
         name: workerName,
-        script: workerScript,
+        script: `
+          addEventListener('fetch', event => {
+            event.respondWith(new Response('Hello world!', { status: 200 }));
+          });
+        `,
         format: "cjs",
       });
 
@@ -307,7 +84,13 @@ describe("Worker Resource", () => {
       // Create a worker with ESM format
       worker = await Worker(workerName, {
         name: workerName,
-        script: esmWorkerScript,
+        script: `
+          export default {
+            async fetch(request, env, ctx) {
+              return new Response('Hello ESM world!', { status: 200 });
+            }
+          };
+        `,
         format: "esm", // Explicitly using ESM
       });
 
@@ -346,7 +129,13 @@ describe("Worker Resource", () => {
       // First create with ESM format
       worker = await Worker(workerName, {
         name: workerName,
-        script: esmWorkerScript,
+        script: `
+          export default {
+            async fetch(request, env, ctx) {
+              return new Response('Hello ESM world!', { status: 200 });
+            }
+          };
+        `,
         format: "esm",
       });
 
@@ -355,7 +144,11 @@ describe("Worker Resource", () => {
       // Update to CJS format
       worker = await Worker(workerName, {
         name: workerName,
-        script: workerScript,
+        script: `
+          addEventListener('fetch', event => {
+            event.respondWith(new Response('Hello world!', { status: 200 }));
+          });
+        `,
         format: "cjs",
       });
       expect(worker.format).toEqual("cjs");
@@ -363,7 +156,13 @@ describe("Worker Resource", () => {
       // Update back to ESM format
       worker = await Worker(workerName, {
         name: workerName,
-        script: esmWorkerScript,
+        script: `
+          export default {
+            async fetch(request, env, ctx) {
+              return new Response('Hello ESM world!', { status: 200 });
+            }
+          };
+        `,
         format: "esm",
       });
 
@@ -381,14 +180,22 @@ describe("Worker Resource", () => {
       // First, create a worker successfully
       await Worker(workerName, {
         name: workerName,
-        script: workerScript,
+        script: `
+          addEventListener('fetch', event => {
+            event.respondWith(new Response('Hello world!', { status: 200 }));
+          });
+        `,
         format: "cjs",
       });
 
       // Try to create another worker with the same name, which should fail
       const duplicateWorker = Worker(`${workerName}-dup`, {
         name: workerName, // Same name as firstWorker
-        script: workerScript,
+        script: `
+          addEventListener('fetch', event => {
+            event.respondWith(new Response('Hello world!', { status: 200 }));
+          });
+        `,
         format: "cjs",
       });
       await expect(duplicateWorker).rejects.toThrow(
@@ -399,88 +206,48 @@ describe("Worker Resource", () => {
     }
   });
 
-  test("create and delete worker with Durable Object binding", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-test-worker-do-binding-do-1`;
-
-    let worker: Worker | undefined;
-    try {
-      // First create the worker without the DO binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: durableObjectWorkerScript,
-        format: "esm",
-        // No bindings yet
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.bindings).toBeEmpty();
-
-      // Create a Durable Object namespace
-      const counterNamespace = new DurableObjectNamespace(
-        "test-counter-namespace",
-        {
-          className: "Counter",
-          scriptName: workerName,
-        },
-      );
-
-      // Update the worker with the DO binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: durableObjectWorkerScript,
-        format: "esm",
-        bindings: {
-          COUNTER: counterNamespace,
-        },
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.bindings).toBeDefined();
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("create and delete worker with KV Namespace binding", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-test-worker-kv-binding-kv-1`;
-
-    let worker: Worker | undefined;
-    let testKv: KVNamespace | undefined;
-    try {
-      // Create a KV namespace with initial values
-      testKv = await KVNamespace("test-kv-namespace", {
-        title: `${BRANCH_PREFIX} Test KV Namespace 2`,
-        values: [
-          {
-            key: "testKey",
-            value: "initial-value",
-          },
-        ],
-      });
-
-      // Create a worker with the KV Namespace binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: kvWorkerScript,
-        format: "esm",
-        bindings: {
-          TEST_KV: testKv,
-        },
-      });
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.bindings).toBeDefined();
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
   test("create and delete worker with multiple bindings", async (scope) => {
     const workerName = `${BRANCH_PREFIX}-test-worker-multi-bindings-multi-1`;
+
+    // Sample ESM worker script with multiple bindings
+    const multiBindingsWorkerScript = `
+  export class Counter {
+    constructor(state, env) {
+      this.state = state;
+      this.env = env;
+      this.counter = 0;
+    }
+
+    async fetch(request) {
+      this.counter++;
+      return new Response('Counter: ' + this.counter, { status: 200 });
+    }
+  }
+
+  export default {
+    async fetch(request, env, ctx) {
+      // Path-based routing to demonstrate different bindings
+      const url = new URL(request.url);
+
+      if (url.pathname.includes('/counter')) {
+        const id = env.COUNTER.idFromName('default');
+        const stub = env.COUNTER.get(id);
+        return stub.fetch(request);
+      }
+
+      if (url.pathname.includes('/kv')) {
+        const value = await env.TEST_KV.get('testKey');
+        return new Response('KV Value: ' + (value || 'not found'), { status: 200 });
+      }
+
+      if (url.pathname.includes('/secret')) {
+        return new Response('Secret: ' + env.API_KEY, { status: 200 });
+      }
+
+      return new Response('Hello worker with multiple bindings!', { status: 200 });
+    }
+  };
+`;
 
     // Create a Durable Object namespace
     const counterNamespace = new DurableObjectNamespace(
@@ -539,6 +306,39 @@ describe("Worker Resource", () => {
   // Add a new test for environment variables
   test("create and test worker with environment variables", async (scope) => {
     const workerName = `${BRANCH_PREFIX}-test-worker-env-vars-env-1`;
+    // Sample ESM worker script with environment variables
+    const envVarsWorkerScript = `
+      export default {
+        async fetch(request, env, ctx) {
+          const url = new URL(request.url);
+
+          // Return the value of the requested environment variable
+          if (url.pathname.startsWith('/env/')) {
+            const varName = url.pathname.split('/env/')[1];
+            const value = env[varName];
+            return new Response(value || 'undefined', {
+              status: 200,
+              headers: { 'Content-Type': 'text/plain' }
+            });
+          }
+
+          // Return all environment variables
+          if (url.pathname === '/env') {
+            const envVars = Object.entries(env)
+              .filter(([key]) => key !== 'COUNTER' && !key.includes('Durable')) // Filter out bindings
+              .map(([key, value]) => \`\${key}: \${value}\`)
+              .join('\\n');
+
+            return new Response(envVars, {
+              status: 200,
+              headers: { 'Content-Type': 'text/plain' }
+            });
+          }
+
+          return new Response('Hello with environment variables!', { status: 200 });
+        }
+      };
+    `;
     let worker: Worker | undefined;
     try {
       // Create a worker with environment variables
@@ -617,177 +417,6 @@ describe("Worker Resource", () => {
     } finally {
       await destroy(scope);
       // Verify the worker was deleted
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("migrate durable object by renaming class", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-test-worker-do-migration-migrate-1`;
-    let worker: Worker | undefined;
-    try {
-      // First create the worker with the original Counter class
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: doMigrationWorkerScriptV1,
-        format: "esm",
-        adopt: true,
-      });
-
-      // Apply to create the worker first
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-
-      // Create a stable DO namespace with the original Counter class
-      const counterNamespace = new DurableObjectNamespace(
-        "test-counter-namespace",
-        {
-          className: "Counter",
-          scriptName: workerName,
-        },
-      );
-
-      // Update worker with the original Counter binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: doMigrationWorkerScriptV1,
-        format: "esm",
-        bindings: {
-          COUNTER: counterNamespace,
-        },
-      });
-
-      expect(worker.bindings).toBeDefined();
-
-      // Now update the namespace to use CounterV2 class
-      const updatedNamespace = new DurableObjectNamespace(
-        "test-counter-namespace",
-        {
-          className: "CounterV2",
-          scriptName: workerName,
-        },
-      );
-
-      // Update worker with the migrated binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: doMigrationWorkerScriptV2,
-        format: "esm",
-        bindings: {
-          COUNTER: updatedNamespace,
-        },
-      });
-
-      expect(worker.bindings).toBeDefined();
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("add environment variables to worker with durable object", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-test-worker-do-with-env-doenv-1`;
-
-    let worker: Worker | undefined;
-    try {
-      // First create a worker with a Durable Object but no env vars
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: durableObjectWorkerScript,
-        format: "esm",
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-
-      // Create a Durable Object namespace
-      const counterNamespace = new DurableObjectNamespace(
-        "test-counter-env-namespace",
-        {
-          className: "Counter",
-          scriptName: workerName,
-        },
-      );
-
-      // Update the worker with the DO binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: durableObjectWorkerScript,
-        format: "esm",
-        bindings: {
-          COUNTER: counterNamespace,
-        },
-      });
-
-      // Apply the worker with binding
-      expect(worker.bindings).toBeDefined();
-      expect(worker.env).toBeUndefined();
-
-      // Now update the worker by adding environment variables
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: durableObjectWorkerScript,
-        format: "esm",
-        bindings: {
-          COUNTER: counterNamespace,
-        },
-        env: {
-          API_SECRET: "test-secret-123",
-          DEBUG_MODE: "true",
-        },
-      });
-
-      expect(worker.bindings).toBeDefined();
-      expect(worker.env).toBeDefined();
-      expect(worker.env?.API_SECRET).toEqual("test-secret-123");
-      expect(worker.env?.DEBUG_MODE).toEqual("true");
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("create and delete worker with R2 bucket binding", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-test-worker-r2-binding-r2-1`;
-
-    // Create a test R2 bucket
-    let testBucket: R2Bucket | undefined;
-
-    let worker: Worker<{ STORAGE: R2Bucket }> | undefined;
-
-    try {
-      testBucket = await R2Bucket("test-bucket", {
-        name: `${BRANCH_PREFIX.toLowerCase()}-test-r2-bucket`,
-        allowPublicAccess: false,
-      });
-
-      // Create a worker with the R2 bucket binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: r2WorkerScript,
-        format: "esm",
-        url: true, // Enable workers.dev URL to test the worker
-        bindings: {
-          STORAGE: testBucket,
-        },
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.bindings).toBeDefined();
-      expect(worker.bindings!.STORAGE).toBeDefined();
-
-      // Test that the R2 binding is accessible in the worker
-      const response = await fetch(`${worker.url}/r2-info`);
-      expect(response.status).toEqual(200);
-      const data = (await response.json()) as {
-        hasR2: boolean;
-        bucketName: string;
-      };
-      expect(data.hasR2).toEqual(true);
-    } finally {
-      await destroy(scope);
       await assertWorkerDoesNotExist(workerName);
     }
   });
@@ -1102,446 +731,6 @@ describe("Worker Resource", () => {
     }
   });
 
-  // Test for binding a workflow to a worker
-  test("create and delete worker with workflow binding", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-test-worker-workflow`;
-
-    // Sample worker script with workflow handler - updated to match Cloudflare Workflows pattern
-    const workflowWorkerScript = `
-      // Workflow definition for email notifications
-      export class EmailNotifier {
-        constructor(state, env) {
-          this.state = state;
-          this.env = env;
-        }
-
-        async run(event, step) {
-          // Process order data from event payload
-          const orderDetails = await step.do('process-order', async () => {
-            console.log("Processing order", event.payload);
-            return {
-              success: true,
-              orderId: event.payload.orderId,
-              message: "Order processed successfully"
-            };
-          });
-
-          return orderDetails;
-        }
-      }
-
-      // Workflow definition for order processing
-      export class OrderProcessor {
-        constructor(state, env) {
-          this.state = state;
-          this.env = env;
-        }
-
-        async run(event, step) {
-          // Process shipping data
-          const shippingDetails = await step.do('process-shipping', async () => {
-            console.log("Processing shipping", event.payload);
-            return {
-              success: true,
-              shipmentId: event.payload.shipmentId,
-              message: "Shipment scheduled successfully"
-            };
-          });
-
-          return shippingDetails;
-        }
-      }
-
-      export default {
-        async fetch(request, env, ctx) {
-          const url = new URL(request.url);
-
-          // Add endpoints to trigger workflows for testing
-          if (url.pathname === '/trigger-email-workflow') {
-            try {
-              // Get workflow binding
-              const workflow = env.EMAIL_WORKFLOW;
-
-              if (!workflow) {
-                return new Response(JSON.stringify({ error: "No email workflow binding found" }), {
-                  status: 500,
-                  headers: { 'Content-Type': 'application/json' }
-                });
-              }
-
-              // Create a workflow instance with parameters
-              const params = { orderId: "test-123", amount: 99.99 };
-              const instance = await workflow.create(params);
-
-              return Response.json({
-                id: instance.id,
-                details: await instance.status(),
-                success: true,
-                orderId: params.orderId,
-                message: "Order processed successfully"
-              });
-            } catch (error) {
-              console.error("Error triggering email workflow:", error);
-              return new Response(JSON.stringify({ error: error.message || "Unknown error" }), {
-                status: 500,
-                headers: { 'Content-Type': 'application/json' }
-              });
-            }
-          }
-
-          // Endpoint for the order workflow
-          if (url.pathname === '/trigger-order-workflow') {
-            try {
-              // Get workflow binding
-              const workflow = env.ORDER_WORKFLOW;
-
-              if (!workflow) {
-                return new Response(JSON.stringify({ error: "No order workflow binding found" }), {
-                  status: 500,
-                  headers: { 'Content-Type': 'application/json' }
-                });
-              }
-
-              // Create a workflow instance with parameters
-              const params = { shipmentId: "ship-456", carrier: "FastShip" };
-              const instance = await workflow.create(params);
-
-              return Response.json({
-                id: instance.id,
-                details: await instance.status(),
-                success: true,
-                shipmentId: params.shipmentId,
-                message: "Shipment scheduled successfully"
-              });
-            } catch (error) {
-              console.error("Error triggering order workflow:", error);
-              return new Response(JSON.stringify({ error: error.message || "Unknown error" }), {
-                status: 500,
-                headers: { 'Content-Type': 'application/json' }
-              });
-            }
-          }
-
-          return new Response('Worker with workflow bindings!', { status: 200 });
-        }
-      };
-    `;
-
-    let worker: Worker | undefined;
-    try {
-      // Create a workflow instance
-      const emailWorkflow = new Workflow("email-notifier", {
-        className: "EmailNotifier",
-        workflowName: "email-notification-workflow",
-      });
-
-      // Create a worker with the workflow binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: workflowWorkerScript,
-        format: "esm",
-        bindings: {
-          EMAIL_WORKFLOW: emailWorkflow,
-        },
-        url: true, // Enable workers.dev URL to test the workflow
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.bindings).toBeDefined();
-      expect(worker.url).toBeTruthy();
-
-      // Test triggering the first workflow
-      const response = await fetch(`${worker.url!}/trigger-email-workflow`);
-      const result: any = await response.json();
-      console.log("Email workflow response:", result);
-
-      expect(response.status).toEqual(200);
-      expect(result.success).toEqual(true);
-      expect(result.orderId).toEqual("test-123");
-      expect(result.message).toEqual("Order processed successfully");
-      // Verify the instance ID is not empty
-      expect(result.id).toBeTruthy();
-      expect(typeof result.id).toBe("string");
-      expect(result.id.length).toBeGreaterThan(0);
-      // Verify the details contain valid status
-      expect(result.details).toBeDefined();
-      expect(result.details.status).toBeTruthy();
-
-      // Create a new workflow binding and update the worker
-      const orderWorkflow = new Workflow("order-processor", {
-        className: "OrderProcessor",
-        workflowName: "order-processing-workflow",
-      });
-
-      // Update the worker with multiple workflow bindings
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: workflowWorkerScript,
-        format: "esm",
-        bindings: {
-          EMAIL_WORKFLOW: emailWorkflow,
-          ORDER_WORKFLOW: orderWorkflow,
-        },
-        url: true,
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(worker.bindings).toBeDefined();
-      expect(Object.keys(worker.bindings || {})).toHaveLength(2);
-
-      // Test triggering the second workflow
-      const orderResponse = await fetch(
-        `${worker.url!}/trigger-order-workflow`,
-      );
-      const orderResult: any = await orderResponse.json();
-      console.log("Order workflow response:", orderResult);
-
-      expect(orderResponse.status).toEqual(200);
-      expect(orderResult.success).toEqual(true);
-      expect(orderResult.shipmentId).toEqual("ship-456");
-      expect(orderResult.message).toEqual("Shipment scheduled successfully");
-      // Verify the instance ID is not empty
-      expect(orderResult.id).toBeTruthy();
-      expect(typeof orderResult.id).toBe("string");
-      expect(orderResult.id.length).toBeGreaterThan(0);
-      // Verify the details contain valid status
-      expect(orderResult.details).toBeDefined();
-      expect(orderResult.details.status).toBeTruthy();
-    } finally {
-      // Explicitly destroy resources since destroy: false is set
-      await destroy(scope);
-      // Verify the worker was deleted
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("create and test worker with D1 database binding", async (scope) => {
-    // Sample ESM worker script with D1 database functionality
-    const d1WorkerScript = `
-      export default {
-        async fetch(request, env, ctx) {
-          const url = new URL(request.url);
-
-          // Initialize the database with a table and data
-          if (url.pathname === '/init-db') {
-            try {
-              const db = env.DATABASE;
-
-              // Create a test table
-              await db.exec("CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY, name TEXT, email TEXT)");
-
-              // Insert some test data
-              await db.exec("INSERT INTO users (name, email) VALUES ('Test User', 'test@example.com')");
-
-              return new Response('Database initialized successfully!', {
-                status: 200,
-                headers: { 'Content-Type': 'text/plain' }
-              });
-            } catch (error) {
-              return new Response('Error initializing database: ' + error.message, {
-                status: 500,
-                headers: { 'Content-Type': 'text/plain' }
-              });
-            }
-          }
-
-          // Query data from the database
-          if (url.pathname === '/query-db') {
-            try {
-              const db = env.DATABASE;
-
-              // Query the database
-              const { results } = await db.prepare("SELECT * FROM users").all();
-
-              return new Response(JSON.stringify({ success: true, data: results }), {
-                status: 200,
-                headers: { 'Content-Type': 'application/json' }
-              });
-            } catch (error) {
-              return new Response(JSON.stringify({
-                success: false,
-                error: error.message
-              }), {
-                status: 500,
-                headers: { 'Content-Type': 'application/json' }
-              });
-            }
-          }
-
-          return new Response('D1 Database Worker is running!', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-      };
-    `;
-
-    const workerName = `${BRANCH_PREFIX}-test-worker-d1`;
-
-    let worker: Worker<{ DATABASE: D1Database }> | undefined;
-    let db: D1Database | undefined;
-
-    try {
-      // Create a D1 database
-      db = await D1Database(`${BRANCH_PREFIX}-test-db`, {
-        name: `${BRANCH_PREFIX}-test-db`,
-        primaryLocationHint: "wnam", // West North America
-      });
-
-      expect(db.id).toBeTruthy();
-      expect(db.name).toEqual(`${BRANCH_PREFIX}-test-db`);
-
-      // Create a worker with the D1 database binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: d1WorkerScript,
-        format: "esm",
-        url: true, // Enable workers.dev URL to test the worker
-        bindings: {
-          DATABASE: db,
-        },
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.bindings).toBeDefined();
-      expect(worker.bindings!.DATABASE).toBeDefined();
-      expect(worker.bindings!.DATABASE.id).toEqual(db.id);
-      expect(worker.url).toBeTruthy();
-
-      // Initialize the database with a table and data
-      const initResponse = await fetch(`${worker.url}/init-db`);
-      expect(initResponse.status).toEqual(200);
-      const initText = await initResponse.text();
-      expect(initText).toEqual("Database initialized successfully!");
-
-      // Query data from the database
-      const queryResponse = await fetch(`${worker.url}/query-db`);
-      expect(queryResponse.status).toEqual(200);
-      const queryData: any = await queryResponse.json();
-      expect(queryData.success).toEqual(true);
-      expect(queryData.data).toBeArray();
-      expect(queryData.data.length).toBeGreaterThan(0);
-      expect(queryData.data[0].name).toEqual("Test User");
-      expect(queryData.data[0].email).toEqual("test@example.com");
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  }, 120000); // Increased timeout for D1 database operations
-
-  test("create and test worker with Queue binding", async (scope) => {
-    // Sample ESM worker script with Queue functionality
-    const queueWorkerScript = `
-      export default {
-        async fetch(request, env, ctx) {
-          const url = new URL(request.url);
-
-          // Send a message to the queue
-          if (url.pathname === '/send-message') {
-            try {
-              const body = await request.json();
-              const messageId = await env.MESSAGE_QUEUE.send(body);
-
-              return new Response(JSON.stringify({
-                success: true,
-                messageId,
-                message: 'Message sent successfully'
-              }), {
-                status: 200,
-                headers: { 'Content-Type': 'application/json' }
-              });
-            } catch (error) {
-              return new Response(JSON.stringify({
-                success: false,
-                error: error.message
-              }), {
-                status: 500,
-                headers: { 'Content-Type': 'application/json' }
-              });
-            }
-          }
-
-          return new Response('Queue Worker is running!', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-      };
-    `;
-
-    const workerName = `${BRANCH_PREFIX}-test-worker-queue`;
-    const queueName = `${BRANCH_PREFIX}-test-queue`;
-
-    let worker: Worker<{ MESSAGE_QUEUE: Queue }> | undefined;
-    let queue: Queue | undefined;
-
-    try {
-      // Create a Queue
-      queue = await Queue(queueName, {
-        name: queueName,
-        settings: {
-          deliveryDelay: 0, // No delay for testing
-          deliveryPaused: false,
-        },
-      });
-
-      expect(queue.id).toBeTruthy();
-      expect(queue.name).toEqual(queueName);
-      expect(queue.type).toEqual("queue");
-
-      // Create a worker with the Queue binding
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: queueWorkerScript,
-        format: "esm",
-        url: true, // Enable workers.dev URL to test the worker
-        bindings: {
-          MESSAGE_QUEUE: queue,
-        },
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.bindings).toBeDefined();
-      expect(worker.bindings!.MESSAGE_QUEUE).toBeDefined();
-      expect(worker.url).toBeTruthy();
-
-      if (worker.url) {
-        // Send a message to the queue
-        const testMessage = {
-          id: "msg-123",
-          content: "Test message content",
-          timestamp: Date.now(),
-        };
-
-        const sendResponse = await fetch(`${worker.url}/send-message`, {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(testMessage),
-        });
-
-        expect(sendResponse.status).toEqual(200);
-        const responseData: any = await sendResponse.json();
-        expect(responseData.success).toEqual(true);
-        expect(responseData.message).toEqual("Message sent successfully");
-      }
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  }, 120000); // Increased timeout for Queue operations
-
   test("create and test worker with Self binding", async (scope) => {
     // Sample ESM worker script with Self binding functionality
 
@@ -1758,6 +947,20 @@ describe("Worker Resource", () => {
 
   test("create and test worker with cron triggers", async (scope) => {
     const workerName = `${BRANCH_PREFIX}-test-worker-cron`;
+
+    // Sample worker script with a scheduled handler
+    const cronWorkerScript = `
+      export default {
+        async fetch(request, env, ctx) {
+          return new Response('Worker with cron is running!', { status: 200 });
+        },
+        async scheduled(event, env, ctx) {
+          // Log the scheduled event details
+          console.log('Scheduled event received:', event.scheduledTime, event.cron);
+          // In a real worker, you would perform tasks here
+        },
+      };
+    `;
 
     let worker: Worker | undefined;
     try {
@@ -2002,53 +1205,6 @@ describe("Worker Resource", () => {
   test("create and delete worker with Analytics Engine binding", async (scope) => {
     const workerName = `${BRANCH_PREFIX}-test-worker-analytics-engine`;
 
-    // Sample worker script that uses Analytics Engine
-    const analyticsWorkerScript = `
-      export default {
-        async fetch(request, env, ctx) {
-          const url = new URL(request.url);
-          
-          // Log an event to the analytics engine
-          if (url.pathname === '/log-event') {
-            try {
-              const body = await request.json();
-              
-              // Write an event to the analytics dataset
-              env.ANALYTICS.writeDataPoint({
-                blobs: [body.action, body.category, body.details || ""],
-                doubles: [body.value || 1.0],
-                indexes: [body.userId || "anonymous"]
-              });
-              
-              return Response.json({
-                success: true,
-                message: "Event logged successfully"
-              });
-            } catch (error) {
-              return Response.json({
-                success: false,
-                error: error.message || "Unknown error"
-              }, { status: 500 });
-            }
-          }
-          
-          // Confirm binding exists
-          if (url.pathname === '/check-binding') {
-            return Response.json({
-              hasBinding: !!env.ANALYTICS,
-              bindingType: typeof env.ANALYTICS,
-              success: true
-            });
-          }
-          
-          return new Response('Analytics Engine Worker is running!', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-      };
-    `;
-
     let worker: Worker | undefined;
     let dataset: AnalyticsEngineDataset | undefined;
 
@@ -2066,7 +1222,51 @@ describe("Worker Resource", () => {
       // Create a worker with the analytics engine binding
       worker = await Worker(workerName, {
         name: workerName,
-        script: analyticsWorkerScript,
+        script: `
+          export default {
+            async fetch(request, env, ctx) {
+              const url = new URL(request.url);
+              
+              // Log an event to the analytics engine
+              if (url.pathname === '/log-event') {
+                try {
+                  const body = await request.json();
+                  
+                  // Write an event to the analytics dataset
+                  env.ANALYTICS.writeDataPoint({
+                    blobs: [body.action, body.category, body.details || ""],
+                    doubles: [body.value || 1.0],
+                    indexes: [body.userId || "anonymous"]
+                  });
+                  
+                  return Response.json({
+                    success: true,
+                    message: "Event logged successfully"
+                  });
+                } catch (error) {
+                  return Response.json({
+                    success: false,
+                    error: error.message || "Unknown error"
+                  }, { status: 500 });
+                }
+              }
+              
+              // Confirm binding exists
+              if (url.pathname === '/check-binding') {
+                return Response.json({
+                  hasBinding: !!env.ANALYTICS,
+                  bindingType: typeof env.ANALYTICS,
+                  success: true
+                });
+              }
+              
+              return new Response('Analytics Engine Worker is running!', {
+                status: 200,
+                headers: { 'Content-Type': 'text/plain' }
+              });
+            }
+          };
+        `,
         format: "esm",
         url: true, // Enable workers.dev URL to test the worker
         bindings: {
@@ -2113,497 +1313,4 @@ describe("Worker Resource", () => {
       await assertWorkerDoesNotExist(workerName);
     }
   }, 60000); // Increase timeout for Worker operations
-
-  test("create and test worker with cross-script durable object binding", async (scope) => {
-    // Create names for both workers
-    const doWorkerName = `${BRANCH_PREFIX}-do-provider-worker`;
-    const clientWorkerName = `${BRANCH_PREFIX}-do-client-worker`;
-
-    // Script for the worker that defines the durable object
-    const doProviderWorkerScript = `
-      export class SharedCounter {
-        constructor(state, env) {
-          this.state = state;
-          this.env = env;
-          this.counter = 0;
-        }
-
-        async fetch(request) {
-          const url = new URL(request.url);
-          
-          if (url.pathname === '/increment') {
-            this.counter++;
-            return new Response(JSON.stringify({
-              action: 'increment',
-              counter: this.counter,
-              worker: '${doWorkerName}'
-            }), {
-              headers: { 'Content-Type': 'application/json' }
-            });
-          }
-
-          if (url.pathname === '/get') {
-            return new Response(JSON.stringify({
-              action: 'get',
-              counter: this.counter,
-              worker: '${doWorkerName}'
-            }), {
-              headers: { 'Content-Type': 'application/json' }
-            });
-          }
-
-          return new Response('SharedCounter DO is running!', { status: 200 });
-        }
-      }
-
-      export default {
-        async fetch(request, env, ctx) {
-          return new Response('DO Provider Worker is running!', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-      };
-    `;
-
-    // Script for the worker that uses the cross-script durable object
-    const clientWorkerScript = `
-      export default {
-        async fetch(request, env, ctx) {
-          const url = new URL(request.url);
-          
-          if (url.pathname === '/increment') {
-            try {
-              // Get the durable object instance and increment
-              const id = env.SHARED_COUNTER.idFromName('test-counter');
-              const stub = env.SHARED_COUNTER.get(id);
-              const response = await stub.fetch(new Request('https://example.com/increment'));
-              const data = await response.json();
-              
-              return Response.json({
-                success: true,
-                clientWorker: '${clientWorkerName}',
-                result: data,
-                crossScriptWorking: true
-              });
-            } catch (error) {
-              return Response.json({
-                success: false,
-                error: error.message,
-                crossScriptWorking: false
-              }, { status: 500 });
-            }
-          }
-
-          if (url.pathname === '/get') {
-            try {
-              // Get the current counter value
-              const id = env.SHARED_COUNTER.idFromName('test-counter');
-              const stub = env.SHARED_COUNTER.get(id);
-              const response = await stub.fetch(new Request('https://example.com/get'));
-              const data = await response.json();
-              
-              return Response.json({
-                success: true,
-                clientWorker: '${clientWorkerName}',
-                result: data
-              });
-            } catch (error) {
-              return Response.json({
-                success: false,
-                error: error.message
-              }, { status: 500 });
-            }
-          }
-
-          return new Response('DO Client Worker is running!', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-      };
-    `;
-
-    let doProviderWorker;
-    let clientWorker;
-
-    try {
-      // First create the worker that defines the durable object with its own binding
-      doProviderWorker = await Worker(doWorkerName, {
-        name: doWorkerName,
-        script: doProviderWorkerScript,
-        format: "esm",
-        url: true, // Enable workers.dev URL
-        bindings: {
-          // Create a durable object namespace for the provider worker
-          SHARED_COUNTER: new DurableObjectNamespace(
-            "provider-counter-namespace",
-            {
-              className: "SharedCounter",
-              // No scriptName means it binds to its own script
-            },
-          ), // Bind to its own durable object
-        },
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(doProviderWorker.id).toBeTruthy();
-      expect(doProviderWorker.name).toEqual(doWorkerName);
-      expect(doProviderWorker.url).toBeTruthy();
-
-      // Create the client worker with the cross-script durable object binding
-      clientWorker = await Worker(clientWorkerName, {
-        name: clientWorkerName,
-        script: clientWorkerScript,
-        format: "esm",
-        url: true, // Enable workers.dev URL
-        bindings: {
-          // Create a cross-script durable object namespace that references the first worker
-          SHARED_COUNTER: new DurableObjectNamespace(
-            "cross-script-counter-namespace",
-            {
-              className: "SharedCounter",
-              scriptName: doWorkerName, // This makes it cross-script
-            },
-          ), // Cross-script binding
-        },
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(clientWorker.id).toBeTruthy();
-      expect(clientWorker.name).toEqual(clientWorkerName);
-      expect(clientWorker.url).toBeTruthy();
-      expect(clientWorker.bindings?.SHARED_COUNTER).toBeDefined();
-
-      // Verify the binding configuration
-      const binding = clientWorker.bindings.SHARED_COUNTER;
-      expect(binding.className).toEqual("SharedCounter");
-      expect(binding.scriptName).toEqual(doWorkerName);
-
-      // Test that both workers respond to basic requests
-      const doProviderResponse = await fetch(doProviderWorker.url!);
-      expect(doProviderResponse.status).toEqual(200);
-      const doProviderText = await doProviderResponse.text();
-      expect(doProviderText).toEqual("DO Provider Worker is running!");
-
-      const clientResponse = await fetch(clientWorker.url!);
-      expect(clientResponse.status).toEqual(200);
-      const clientText = await clientResponse.text();
-      expect(clientText).toEqual("DO Client Worker is running!");
-
-      // Test cross-script durable object functionality by calling increment
-      const incrementResponse = await fetch(`${clientWorker.url}/increment`);
-      expect(incrementResponse.status).toEqual(200);
-      const incrementData: any = await incrementResponse.json();
-
-      expect(incrementData).toMatchObject({
-        success: true,
-        clientWorker: clientWorkerName,
-        crossScriptWorking: true,
-        result: {
-          action: "increment",
-          worker: doWorkerName,
-        },
-      });
-
-      // Test getting the counter value
-      const getResponse = await fetch(`${clientWorker.url}/get`);
-      expect(getResponse.status).toEqual(200);
-      const getData: any = await getResponse.json();
-
-      expect(getData).toMatchObject({
-        success: true,
-        clientWorker: clientWorkerName,
-        result: {
-          action: "get",
-          worker: doWorkerName,
-        },
-      });
-
-      // Test state persistence by making multiple increment calls
-      const firstIncrement = await fetch(`${clientWorker.url}/increment`);
-      const firstData: any = await firstIncrement.json();
-
-      const secondIncrement = await fetch(`${clientWorker.url}/increment`);
-      const secondData: any = await secondIncrement.json();
-
-      const thirdIncrement = await fetch(`${clientWorker.url}/increment`);
-      const thirdData: any = await thirdIncrement.json();
-
-      // Verify that the counter is incrementing properly across calls
-      expect(firstData.result.counter).toBeLessThan(secondData.result.counter);
-      expect(secondData.result.counter).toBeLessThan(thirdData.result.counter);
-    } finally {
-      await destroy(scope);
-      // Verify both workers were deleted
-      await assertWorkerDoesNotExist(doWorkerName);
-      await assertWorkerDoesNotExist(clientWorkerName);
-    }
-  }, 120000); // Increased timeout for cross-script DO operations
-
-  test("create and test worker with cross-script durable object binding using re-exported syntax", async (scope) => {
-    // Create names for both workers
-    const doWorkerName = `${BRANCH_PREFIX}-do-provider-worker-re-exported`;
-    const clientWorkerName = `${BRANCH_PREFIX}-do-client-worker-re-exported`;
-
-    // Script for the worker that defines the durable object
-    const doProviderWorkerScript = `
-      export class SharedCounter {
-        constructor(state, env) {
-          this.state = state;
-          this.env = env;
-          this.counter = 0;
-        }
-
-        async fetch(request) {
-          const url = new URL(request.url);
-          
-          if (url.pathname === '/increment') {
-            this.counter++;
-            return new Response(JSON.stringify({
-              action: 'increment',
-              counter: this.counter,
-              worker: '${doWorkerName}'
-            }), {
-              headers: { 'Content-Type': 'application/json' }
-            });
-          }
-
-          if (url.pathname === '/get') {
-            return new Response(JSON.stringify({
-              action: 'get',
-              counter: this.counter,
-              worker: '${doWorkerName}'
-            }), {
-              headers: { 'Content-Type': 'application/json' }
-            });
-          }
-
-          return new Response('SharedCounter DO is running!', { status: 200 });
-        }
-      }
-
-      export default {
-        async fetch(request, env, ctx) {
-          return new Response('DO Provider Worker is running!', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-      };
-    `;
-
-    // Script for the worker that uses the cross-script durable object
-    const clientWorkerScript = `
-      export default {
-        async fetch(request, env, ctx) {
-          const url = new URL(request.url);
-          
-          if (url.pathname === '/increment') {
-            try {
-              // Get the durable object instance and increment
-              const id = env.SHARED_COUNTER.idFromName('test-counter');
-              const stub = env.SHARED_COUNTER.get(id);
-              const response = await stub.fetch(new Request('https://example.com/increment'));
-              const data = await response.json();
-              
-              return Response.json({
-                success: true,
-                clientWorker: '${clientWorkerName}',
-                result: data,
-                crossScriptWorking: true
-              });
-            } catch (error) {
-              return Response.json({
-                success: false,
-                error: error.message,
-                crossScriptWorking: false
-              }, { status: 500 });
-            }
-          }
-
-          if (url.pathname === '/get') {
-            try {
-              // Get the current counter value
-              const id = env.SHARED_COUNTER.idFromName('test-counter');
-              const stub = env.SHARED_COUNTER.get(id);
-              const response = await stub.fetch(new Request('https://example.com/get'));
-              const data = await response.json();
-              
-              return Response.json({
-                success: true,
-                clientWorker: '${clientWorkerName}',
-                result: data
-              });
-            } catch (error) {
-              return Response.json({
-                success: false,
-                error: error.message
-              }, { status: 500 });
-            }
-          }
-
-          return new Response('DO Client Worker is running!', {
-            status: 200,
-            headers: { 'Content-Type': 'text/plain' }
-          });
-        }
-      };
-    `;
-
-    let doProviderWorker;
-    let clientWorker;
-
-    try {
-      // First create the worker that defines the durable object with its own binding
-      doProviderWorker = await Worker(doWorkerName, {
-        name: doWorkerName,
-        script: doProviderWorkerScript,
-        format: "esm",
-        url: true, // Enable workers.dev URL
-        bindings: {
-          // Create a durable object namespace for the provider worker
-          SHARED_COUNTER: new DurableObjectNamespace(
-            "provider-counter-namespace",
-            {
-              className: "SharedCounter",
-              // No scriptName means it binds to its own script
-            },
-          ), // Bind to its own durable object
-        },
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(doProviderWorker.id).toBeTruthy();
-      expect(doProviderWorker.name).toEqual(doWorkerName);
-      expect(doProviderWorker.url).toBeTruthy();
-
-      expect(doProviderWorker.bindings.SHARED_COUNTER.scriptName).toEqual(
-        doWorkerName,
-      );
-
-      // Create the client worker with the cross-script durable object binding
-      clientWorker = await Worker(clientWorkerName, {
-        name: clientWorkerName,
-        script: clientWorkerScript,
-        format: "esm",
-        url: true, // Enable workers.dev URL
-        bindings: {
-          // Create a cross-script durable object namespace that references the first worker
-          SHARED_COUNTER: doProviderWorker.bindings.SHARED_COUNTER,
-        },
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      expect(clientWorker.id).toBeTruthy();
-      expect(clientWorker.name).toEqual(clientWorkerName);
-      expect(clientWorker.url).toBeTruthy();
-      expect(clientWorker.bindings?.SHARED_COUNTER).toBeDefined();
-
-      // Verify the binding configuration
-      const binding = clientWorker.bindings.SHARED_COUNTER;
-      expect(binding.className).toEqual("SharedCounter");
-      expect(binding.scriptName).toEqual(doWorkerName);
-
-      // Test that both workers respond to basic requests
-      const doProviderResponse = await fetch(doProviderWorker.url!);
-      expect(doProviderResponse.status).toEqual(200);
-      const doProviderText = await doProviderResponse.text();
-      expect(doProviderText).toEqual("DO Provider Worker is running!");
-
-      const clientResponse = await fetch(clientWorker.url!);
-      expect(clientResponse.status).toEqual(200);
-      const clientText = await clientResponse.text();
-      expect(clientText).toEqual("DO Client Worker is running!");
-
-      // Test cross-script durable object functionality by calling increment
-      const incrementResponse = await fetch(`${clientWorker.url}/increment`);
-      expect(incrementResponse.status).toEqual(200);
-      const incrementData: any = await incrementResponse.json();
-
-      expect(incrementData).toMatchObject({
-        success: true,
-        clientWorker: clientWorkerName,
-        crossScriptWorking: true,
-        result: {
-          action: "increment",
-          worker: doWorkerName,
-        },
-      });
-
-      // Test getting the counter value
-      const getResponse = await fetch(`${clientWorker.url}/get`);
-      expect(getResponse.status).toEqual(200);
-      const getData: any = await getResponse.json();
-
-      expect(getData).toMatchObject({
-        success: true,
-        clientWorker: clientWorkerName,
-        result: {
-          action: "get",
-          worker: doWorkerName,
-        },
-      });
-
-      // Test state persistence by making multiple increment calls
-      const firstIncrement = await fetch(`${clientWorker.url}/increment`);
-      const firstData: any = await firstIncrement.json();
-
-      const secondIncrement = await fetch(`${clientWorker.url}/increment`);
-      const secondData: any = await secondIncrement.json();
-
-      const thirdIncrement = await fetch(`${clientWorker.url}/increment`);
-      const thirdData: any = await thirdIncrement.json();
-
-      // Verify that the counter is incrementing properly across calls
-      expect(firstData.result.counter).toBeLessThan(secondData.result.counter);
-      expect(secondData.result.counter).toBeLessThan(thirdData.result.counter);
-    } finally {
-      await destroy(scope);
-      // Verify both workers were deleted
-      await assertWorkerDoesNotExist(doWorkerName);
-      await assertWorkerDoesNotExist(clientWorkerName);
-    }
-  }, 120000); // Increased timeout for cross-script DO operations
-
-  test("adopting a Worker should use server-side state to migrate classes", async (scope) => {
-    try {
-      const workerName = `${BRANCH_PREFIX}-test-worker-adoption-migrate`;
-      await Worker("worker-1", {
-        name: workerName,
-        script: `
-          export class Counter {}
-          export default { fetch() {} }
-        `,
-        bindings: {
-          DO: new DurableObjectNamespace("DO", {
-            className: "Counter",
-          }),
-        },
-      });
-
-      await Worker("worker-2", {
-        name: workerName,
-        // adopt the worker since it already exists
-        adopt: true,
-        script: `
-          export class Counter2 {}
-          export default { fetch() {} }
-        `,
-        bindings: {
-          // mapped by stable ID "DO"
-          DO_1: new DurableObjectNamespace("DO", {
-            // should migrate to Counter 2
-            className: "Counter2",
-          }),
-        },
-      });
-    } finally {
-      await destroy(scope);
-    }
-  });
 });
